+++ lib/CHANGELOG.md
@@ -5,11 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Feat
+- add chart `onDblClick` prop to handle double-click event subscription
 
 ## [0.3.4] - 2025-05-09
 ### Fix
 - fix time scale options not being applied on init
-- feat: add `<TimeScaleFitContentTrigger>` component
+### Feat
+- add `<TimeScaleFitContentTrigger>` component
 
 ## [0.3.3] - 2025-04-21
 ### Feat