+++ lib/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.1] - 2025-06-06
+### Changed
+- migrate to `classic` jsx transform from `react-jsx`
+
 ## [0.4.0] - 2025-06-03
 ### Added
 - add `SeriesPrimitive` component